Melsec Word Data Ex 함수 사용 - PLC 데이터 주소가 short 넘어감

+++ NotchingGradeInsp/NotchingInspPlc/MelsecPlcIo.cpp
@@ -367,7 +367,7 @@ int CMelsecPlcIo::LocalPioOpen(BOOL bLockCtrl)
 
 	int	iRet = 0;
 
-	iRet = (int)mdOpen(m_wChannelNo, -1, &m_pPath);
+	iRet = (int)mdOpen((short)m_wChannelNo, -1, &m_pPath);
 
 	if (iRet == 0) 
 	{
